Attempt to fix wrong images for tiles
